fix config issues

+++ migrations/20231031073421-create-user.js
@@ -23,7 +23,7 @@ module.exports = {
         allowNull: false,
       },
       mobile: {
-        type: Sequelize.INTEGER(10),
+        type: Sequelize.INTEGER,
         allowNull: false,
         unique: true,
       },
